requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
addict=2.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.9.24=haa95532_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
comtypes=1.4.8=pypi_0
contourpy=1.3.0=pypi_0
cpython=3.10.15=py310hd8ed1ab_2
cycler=0.12.1=pypi_0
debugpy=1.8.7=py310h9e98ed7_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
filelock=3.16.1=pypi_0
filterpy=1.4.5=pypi_0
fire=0.7.0=pypi_0
fonttools=4.54.1=pypi_0
fsspec=2024.10.0=pypi_0
gitpy=0.0.5=pypi_0
idna=3.10=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.5.0=pyha770c72_0
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.28.0=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh5737063_1
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=hdf4eb48_0
libffi=3.4.4=hd77b12b_1
libsodium=1.0.20=hc70643c_0
logger=1.4=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mmengine=0.10.5=pypi_0
motpy=0.0.10=pypi_0
mpmath=1.3.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.4.2=pypi_0
numpy=2.1.2=pypi_0
omegaconf=2.3.0=pypi_0
opencv-contrib-python=4.10.0.84=pypi_0
openssl=3.3.2=h2466b09_0
packaging=24.1=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=11.0.0=pypi_0
pip=24.2=py310haa95532_0
platformdirs=4.3.6=pyhd8ed1ab_0
prompt-toolkit=3.0.48=pyha770c72_0
psutil=6.0.0=py310ha8f682b_2
pure_eval=0.2.3=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.2.0=pypi_0
pyrealsense2=2.55.1.6486=pypi_0
python=3.10.15=h4607a30_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=2_cp310
pywin32=307=py310h9e98ed7_3
pywinauto=0.6.8=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py310h656833d_3
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
scipy=1.14.1=pypi_0
screeninfo=0.8.1=pypi_0
setuptools=75.1.0=py310haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.1=pypi_0
termcolor=2.5.0=pypi_0
tk=8.6.14=h0416ee5_0
tomli=2.2.1=pypi_0
torch=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.1=py310ha8f682b_1
tqdm=4.67.0=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_1
unicode=2.9=pypi_0
urllib3=2.2.3=pypi_0
vc=14.40=h2eaa2aa_1
vc14_runtime=14.40.33810=hcc2c482_22
vs2015_runtime=14.40.33810=h3bf8584_22
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.44.0=py310haa95532_0
xz=5.4.6=h8cc25b3_1
yapf=0.43.0=pypi_0
zeromq=4.3.5=ha9f60a1_6
zipp=3.20.2=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_1
